// src/app/layout/broker/models/cliente/cliente.ts
export class Cliente {
  p_NIDPROCESS: string;
  p_NPERSON_TYP: string; // TIPO DE PERSONA
  p_NDOCUMENT_TYP: number; // TIPO DE DOCUMENTO
  p_SDOCUMENT: string; // NUMERO DE DOCUMENTO
  p_SLEGALNAME: string; // RAZON SOCIAL
  p_SCLIENT_NAME: string; // NOMBRE
  p_SCLIENT_APPPAT: string; // APELLIDO PATERNO
  p_SCLIENT_APPMAT: string; // APELLIDO MATERNO
  p_NMUNICIPALITY: number; // DISTRITO
  p_NLOCAT: number; // PROVINCIA
  p_NPROVINCE: number; // DEPARTAMENTO
  p_SADDRESS: string; // DIRECCION
  p_SMAIL: string; // MAIL
  p_SPHONE: string; // TELEFONO
  V_NIDPROCESS: number;
  p_NCODINEI: number; // DISTRITO INEI
  p_SCANAL: string; // TELEFONO
}